UI Card detail: Externalize card acknowlegment to another component
