+++ ui/main/src/app/modules/cards/components/card-ack/card-ack.component.html
@@ -1,0 +1,14 @@
+<!-- Copyright (c) 2022, RTE (http://www.rte-france.com)              -->
+<!-- See AUTHORS.txt                                                       -->
+<!-- This Source Code Form is subject to the terms of the Mozilla Public   -->
+<!-- License, v. 2.0. If a copy of the MPL was not distributed with this   -->
+<!-- file, You can obtain one at http://mozilla.org/MPL/2.0/.              -->
+<!-- SPDX-License-Identifier: MPL-2.0                                      -->
+
+<button id="opfab-card-details-btn-ack" class="opfab-btn" *ngIf='showAckButton'
+(click)='acknowledgeCard()'>{{btnAckText|translate}}</button>
+
+<of-spinner *ngIf="ackOrUnackInProgress"
+[seeInModal]="true"
+loadingText="shared.requestInProgress">
+</of-spinner>
